requirements-dev.txt
mypy==0.812
pycodestyle
pylint==2.4.4
vcrpy
pproxy==2.1.8
Sphinx>=1.8.3
sphinx-rtd-theme>=0.4.0
sphinx-tabs>=1.1.11
sphinx-automodapi
sphinx-autodoc-typehints
matplotlib>=2.1
jupyter
plotly>=4.4
ipyvuetify>=1.1
ipyvue>=1.4.1
ipython>=5.0.0
seaborn>=0.9.0
pyperclip>=1.7
jupyter-sphinx
reno>=2.11.0
scipy>=1.0
nbformat>=4.4.0
nbconvert>=5.3.1
qiskit_rng
qiskit-aer
websockets>=8
scikit-quant;platform_system != 'Windows'
black==22.1.0